+++ src/DriveShare.php
@@ -2,52 +2,59 @@
 
 namespace Owncloud\OcisPhpSdk;
 
-use OpenAPI\Client\Model\Permission as ApiPermission;
+use GuzzleHttp\Client;
+use OpenAPI\Client\Api\DrivesRootApi;
+use OpenAPI\Client\ApiException;
+use Owncloud\OcisPhpSdk\Exception\BadRequestException;
+use Owncloud\OcisPhpSdk\Exception\EndPointNotImplementedException;
+use Owncloud\OcisPhpSdk\Exception\ExceptionHelper;
+use Owncloud\OcisPhpSdk\Exception\ForbiddenException;
+use Owncloud\OcisPhpSdk\Exception\HttpException;
+use Owncloud\OcisPhpSdk\Exception\InternalServerErrorException;
 use Owncloud\OcisPhpSdk\Exception\InvalidResponseException;
+use Owncloud\OcisPhpSdk\Exception\NotFoundException;
+use Owncloud\OcisPhpSdk\Exception\UnauthorizedException;
 
-class DriveShare
+class DriveShare extends Share
 {
-    private ApiPermission $apiPermission;
-    private string $driveId;
+    protected function getDrivesRootApi(): DrivesRootApi
+    {
+        $guzzle = new Client(
+            Ocis::createGuzzleConfig($this->connectionConfig, $this->accessToken)
+        );
+        if (array_key_exists('drivesRootApi', $this->connectionConfig)) {
+            return $this->connectionConfig['drivesRootApi'];
+        }
 
+        return new DrivesRootApi(
+            $guzzle,
+            $this->graphApiConfig
+        );
+    }
 
     /**
-     * @ignore The developer using the SDK does not need to create share objects manually,
-     *         but should use the OcisResource class to invite people to a resource and
-     *         that will create ShareCreated objects
+     * Permanently delete the current share or share link
+     *
+     * @throws BadRequestException
+     * @throws ForbiddenException
+     * @throws HttpException
+     * @throws NotFoundException
+     * @throws UnauthorizedException
+     * @throws InvalidResponseException
+     * @throws InternalServerErrorException
+     * @throws EndPointNotImplementedException
      */
-    public function __construct(
-        ApiPermission $apiPermission,
-        string        $driveId
-    ) {
-        $this->apiPermission = $apiPermission;
-        $this->driveId = $driveId;
-    }
-
-    public function getPermissionId(): string
+    public function delete(): bool
     {
-        $id = $this->apiPermission->getId();
-        if ($id === null || $id === '') {
-            throw new InvalidResponseException(
-                "Invalid id returned for permission '" . print_r($id, true) . "'"
+        try {
+            $this->getDrivesRootApi()->deletePermissionSpaceRoot(
+                $this->driveId,
+                $this->getPermissionId()
             );
+        } catch (ApiException $e) {
+            throw ExceptionHelper::getHttpErrorException($e);
         }
-        return $id;
+        return true;
     }
 
-    public function getExpiration(): ?\DateTimeImmutable
-    {
-        $expiry = $this->apiPermission->getExpirationDateTime();
-        if ($expiry === null) {
-            return null;
-        } else {
-            return \DateTimeImmutable::createFromMutable($expiry);
-        }
-
-    }
-
-    public function getDriveId(): string
-    {
-        return $this->driveId;
-    }
 }
